Add usize and isize types
Arrays can only be indexed by usize and this enforces that rule. These
types are sized based on the pointer size of the host arch.

Fixes #87

+++ gcc/rust/typecheck/rust-hir-type-check-expr.h
@@ -631,7 +631,7 @@ public:
   void visit (HIR::ArrayIndexExpr &expr)
   {
     TyTy::TyBase *size_ty;
-    if (!context->lookup_builtin ("i32", &size_ty))
+    if (!context->lookup_builtin ("usize", &size_ty))
       {
 	rust_error_at (
 	  expr.get_locus (),
